backend/requirements.txt
fastapi
uvicorn
redis
python-dotenv
websockets
typing-extensions
